Move partial classes and methods snippets
