+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideObjects/CS/Objects.cs
@@ -132,147 +132,13 @@ namespace CsCsrefProgrammingObjects
     }
     //</Snippet11>
 
-    namespace WrapCoords2
-    {
-        //<Snippet17>
-        public partial class Coords
-        {
-            private int x;
-            private int y;
+    //-----------------------------------------------------------------------------
 
-            public Coords(int x, int y)
-            {
-                this.x = x;
-                this.y = y;
-            }
-        }
-
-        public partial class Coords
-        {
-            public void PrintCoords()
-            {
-                Console.WriteLine("Coords: {0},{1}", x, y);
-            }
-        }
-
-        class TestCoords
-        {
-            static void Main()
-            {
-                Coords myCoords = new Coords(10, 15);
-                myCoords.PrintCoords();
-
-                // Keep the console window open in debug mode.
-                Console.WriteLine("Press any key to exit.");
-                Console.ReadKey();
-            }
-        }
-        // Output: Coords: 10,15
-        //</Snippet17>
-    }
 
     //-----------------------------------------------------------------------------
-    //<Snippet18>
-    partial interface ITest
-    {
-        void Interface_Test();
-    }
 
-    partial interface ITest
-    {
-        void Interface_Test2();
-    }
 
-    partial struct S1
-    {
-        void Struct_Test() { }
-    }
 
-    partial struct S1
-    {
-        void Struct_Test2() { }
-    }
-    //</Snippet18>
-
-    //-----------------------------------------------------------------------------
-    //<Snippet19>
-    partial class ClassWithNestedClass
-    {
-        partial class NestedClass { }
-    }
-
-    partial class ClassWithNestedClass
-    {
-        partial class NestedClass { }
-    }
-    //</Snippet19>
-
-    //-----------------------------------------------------------------------------
-    //<Snippet20>
-    public partial class A { }
-    //public class A { }  // Error, must also be marked partial
-    //</Snippet20>
-
-    //-----------------------------------------------------------------------------
-    class Planet { }
-    interface IRotate { }
-    interface IRevolve { }
-
-    //<Snippet21>
-    partial class Earth : Planet, IRotate { }
-    partial class Earth : IRevolve { }
-    //</Snippet21>
-
-    //<Snippet23>
-    [SerializableAttribute]
-    partial class Moon { }
-
-    [ObsoleteAttribute]
-    partial class Moon { }
-    //</Snippet23>
-
-    class WrapEquivs
-    {
-        //<Snippet22>
-        class Earth : Planet, IRotate, IRevolve { }
-        //</Snippet22>
-
-        //<Snippet24>
-        [SerializableAttribute]
-        [ObsoleteAttribute]
-        class Moon { }
-        //</Snippet24>
-    }
-
-    //<Snippet25>
-    class Container
-    {
-        partial class Nested
-        {
-            void Test() { }
-        }
-        partial class Nested
-        {
-            void Test2() { }
-        }
-    }
-    //</Snippet25>
-
-    //<Snippet26>
-    public partial class Employee
-    {
-        public void DoWork()
-        {
-        }
-    }
-
-    public partial class Employee
-    {
-        public void GoToLunch()
-        {
-        }
-    }
-    //</Snippet26>
 
     //-----------------------------------------------------------------------------
     //<Snippet27>
